+++ packages/plugin-rax-router/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 1.0.4
+
+- Fix: spa page auto execute
+- Fix: keep alive page error
+
 ## 1.0.3
 
 - Fix: ssr component render